Remove local variable, maxNameLength, in LadderFormat class
Cannot initialize a static variable in LadderFormat using a constructor
because an instance of the class will own the variable, not the class.
So, I figured it would be better to keep passing the variable as a parameter
of LadderBuilder's methods down all the way to where LadderFormat's
static methods are being called.